pyproject.toml
[project]
name = "ironsbot"
version = "0.1.0"
description = "赛尔号信息查询机器人"
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "nonebot-adapter-onebot>=2.4.6",
    "seerapi>=104.0.0",
    "nonebot2[fastapi,httpx]>=2.4.4",
    "seerapi-models>=104.0.0",
    "nonebot-plugin-htmlkit>=0.1.0rc5",
    "hishel[httpx]>=1.1.9",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-treehelp>=0.5.0",
    "nonebot-plugin-send-anything-anywhere>=0.7.1",
]

[project.optional-dependencies]
dev = ["basedpyright", "ruff"]

[tool.nonebot]
plugin_dirs = ["ironsbot/plugins"]
builtin_plugins = []

[tool.nonebot.adapters]
nonebot-adapter-onebot = [
    { name = "OneBot V12", module_name = "nonebot.adapters.onebot.v12" },
]
"@local" = []

[tool.nonebot.plugins]
"@local" = []
nonebot-plugin-htmlkit = ["nonebot_plugin_htmlkit"]
nonebot-plugin-apscheduler = ["nonebot_plugin_apscheduler"]
nonebot-plugin-treehelp = ["nonebot_plugin_treehelp"]
nonebot-plugin-send-anything-anywhere = ["nonebot_plugin_saa"]
[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
# For more rules, see https://docs.astral.sh/ruff/rules/.
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "I",     # isort
    "C90",   # mccabe
    "N",     # pep8-naming
    "PL",    # pylint
    "UP",    # pyupgrade
    "YTT",   # flake8-2020
    "ANN",   # flake8-annotations
    "ASYNC", # flake8-async
    "BLE",   # flake8-blind-except
    "FBT",   # flake8-boolean-trap
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "COM",   # flake8-commas
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "T10",   # flake8-debugger
    "ICN",   # flake8-import-conventions
    "PIE",   # flake8-pie
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "Q",     # flake8-quotes
    "RSE",   # flake8-raise
    "RET",   # flake8-return
    "SIM",   # flake8-simplify
    "SLOT",  # flake8-slots
    "TID",   # flake8-tidy-imports
    "TC",    # flake8-type-checking
    "ARG",   # flake8-unused-arguments
    "PTH",   # flake8-use-pathlib
    # "ERA",   # eradicate
    "FAST", # FastAPI
    "PERF", # Perflint
    "PGH",  # pygrep-hooks
    "FURB", # refurb
    "TRY",  # tryceratops
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E402",  # module-import-not-at-top-of-file  # nonebot2 require() violates this
    "B008",  # function-call-in-default-argument  # nonebot2 Depends() without Annotated violates this
    "UP037", # quoted-annotation
    # "RUF001",  # ambiguous-unicode-character-string
    # "RUF002",  # ambiguous-unicode-character-docstring
    # "RUF003",  # ambiguous-unicode-character-comment
    # "ANN201",  # missing-return-type-undocumented-public-function
    "ANN202",  # missing-return-type-private-function
    "ANN401",  # any-type
    "COM812",  # missing-trailing-comma
    "PLC0415", # import-outside-top-level
]
allowed-confusables = [
    "，",
    "。",
    "“",
    "”",
    "：",
    "；",
    "？",
    "！",
    "【",
    "】",
    "《",
    "》",
    "…",
    "—",
    "（",
    "）",
    "、",
]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.basedpyright]
# For BasedPyright configurations, see https://docs.basedpyright.com/latest/configuration/config-files/.
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "standard"

[dependency-groups]
dev = ["basedpyright>=1.38.1", "ruff>=0.15.2"]
